cleanup + layout structure
