+++ include/utils.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <Arduino.h>
+
+String getMacSuffix();
